# src/config/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Recipe API",
    "version": "1.0.0",
    "description": "API pentru gestionarea rețetelor culinare"
  },
  "servers": [
    {
      "url": "http://localhost:5000"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "Recipe": {
        "type": "object",
        "properties": {
          "id": { "type": "integer" },
          "title": { "type": "string" },
          "description": { "type": "string" },
          "image": { "type": "string" },
          "ingredients": { "type": "string" },
          "instructions": { "type": "string" },
          "prepTime": { "type": "integer" },
          "cookTime": { "type": "integer" },
          "servings": { "type": "integer" },
          "recipeType": { "type": "string" },
          "cuisine": {
            "type": "object",
            "properties": {
              "id": { "type": "integer" },
              "name": { "type": "string" },
              "description": { "type": "string" },
              "image": { "type": "string" }
            }
          },
          "source": { "type": "string" },
          "userId": { "type": "integer", "nullable": true },
          "isAIGenerated": { "type": "boolean" },
          "isPrivate": { "type": "boolean" },
          "createdAt": { "type": "string", "format": "date-time" },
          "updatedAt": { "type": "string", "format": "date-time" }
        }
      },
      "Cuisine": {
        "type": "object",
        "properties": {
          "id": { "type": "integer" },
          "name": { "type": "string" },
          "description": { "type": "string" },
          "image": { "type": "string" }
        }
      },
...
    }
  },
  "paths": {
    "/api/recipes": {
      "get": {
        "summary": "Preia toate rețetele",
        "description": "Preia lista tuturor rețetelor din baza de date",
        "responses": {
          "200": {
            "description": "Lista tuturor rețetelor",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/Recipe" }
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Adaugă o rețetă nouă",
        "description": "Creează o rețetă nouă în baza de date",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "title": { "type": "string" },
                  "description": { "type": "string" },
                  "image": { "type": "string" },
                  "ingredients": { "type": "string" },
                  "instructions": { "type": "string" },
                  "prepTime": { "type": "integer" },
                  "cookTime": { "type": "integer" },
                  "servings": { "type": "integer" },
                  "recipeType": { "type": "string" },
                  "cuisineId": { "type": "integer" },
                  "source": { "type": "string" },
                  "userId": { "type": "integer", "nullable": true },
                  "isAIGenerated": { "type": "boolean" },
                  "isPrivate": { "type": "boolean" }
                },
                "required": ["title", "description", "ingredients", "instructions", "cuisineId"]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Rețeta a fost creată cu succes",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Recipe" }
              }
            }
          },
          "400": { "description": "Date invalide" },
          "500": { "description": "Eroare server" }
        }
      }
    },
    "/api/recipes/{id}": {
      "get": {
        "summary": "Preia o rețetă după ID",
        "description": "Returnează detaliile unei rețete specifice, pe baza ID-ului.",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": { "type": "integer" },
            "description": "ID-ul rețetei"
          }
        ],
        "responses": {
          "200": {
            "description": "Detaliile rețetei",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Recipe" }
              }
            }
          },
          "404": { "description": "Rețetă negăsită" },
          "500": { "description": "Eroare server" }
        }
      },
      "patch": {
        "summary": "Actualizează o rețetă existentă",
        "description": "Actualizează parțial o rețetă existentă",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": { "type": "integer" },
            "description": "ID-ul rețetei"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "title": { "type": "string" },
                  "description": { "type": "string" },
                  "image": { "type": "string" },
                  "ingredients": { "type": "string" },
                  "instructions": { "type": "string" },
                  "prepTime": { "type": "integer" },
                  "cookTime": { "type": "integer" },
                  "servings": { "type": "integer" },
                  "recipeType": { "type": "string" },
                  "cuisineId": { "type": "integer" },
                  "source": { "type": "string" },
                  "isPrivate": { "type": "boolean" }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Rețeta a fost actualizată cu succes",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Recipe" }
              }
            }
          },
          "404": { "description": "Rețetă negăsită" },
          "500": { "description": "Eroare server" }
        }
      },
      "delete": {
        "summary": "Șterge o rețetă",
        "description": "Șterge o rețetă din baza de date",
        "parameters": [
  {
    "in": "path",
    "name": "id",
    "required": true,
    "schema": { "type": "integer" },
    "description": "ID-ul rețetei"
  },
  {
    "in": "query",
    "name": "userId",
    "required": true,
    "schema": { "type": "integer" },
    "description": "ID-ul utilizatorului care șterge rețeta"
  }
],
        "responses": {
          "200": { "description": "Rețeta a fost ștearsă cu succes" },
          "403": { "description": "Nu ai permisiunea să ștergi această rețetă" },
          "404": { "description": "Rețetă negăsită" },
          "500": { "description": "Eroare server" }
        }
      }
    },
    "/api/recipes/sorted": {
      "get": {
        "summary": "Preia rețetele sortate după timpul total",
        "description": "Returnează rețetele sortate după timpul total (prepTime + cookTime)",
        "responses": {
          "200": {
            "description": "Lista rețetelor sortate",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/Recipe" }
                }
              }
            }
          },
          "500": { "description": "Eroare server" }
        }
      }
    },
    "/api/recipes/type/{type}": {
      "get": {
        "summary": "Preia rețetele după tip",
        "description": "Filtrarea rețetelor după tipul lor (mic-dejun, prânz, etc.)",
        "parameters": [
          {
            "in": "path",
            "name": "type",
            "required": true,
            "schema": { "type": "string" },
            "description": "Tipul rețetei (ex: mic-dejun, prânz, etc.)"
          }
        ],
        "responses": {
          "200": {
            "description": "Lista rețetelor după tip",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/Recipe" }
                }
              }
            }
          }
        }
      }
    },
    "/api/recipes/user/{userId}": {
      "get": {
        "summary": "Preia rețetele utilizatorului",
        "description": "Obține rețetele adăugate de un utilizator specific (nu AI generate)",
        "parameters": [
          {
            "in": "path",
            "name": "userId",
            "required": true,
            "schema": { "type": "integer" },
            "description": "ID-ul utilizatorului"
          }
        ],
        "responses": {
          "200": {
            "description": "Lista rețetelor utilizatorului",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/Recipe" }
                }
              }
            }
          },
          "500": { "description": "Eroare server" }
        }
      }
    },
    "/api/recipes/ai/{userId}": {
      "get": {
        "summary": "Preia rețetele AI generate de utilizator",
        "description": "Obține rețetele AI generate de un utilizator specific",
        "parameters": [
          {
            "in": "path",
            "name": "userId",
            "required": true,
            "schema": { "type": "integer" },
            "description": "ID-ul utilizatorului"
          }
        ],
        "responses": {
          "200": {
            "description": "Lista rețetelor AI generate",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/Recipe" }
                }
              }
            }
          },
          "500": { "description": "Eroare server" }
        }
      }
    },
    "/api/recipes/search/{query}": {
      "get": {
        "summary": "Căutare rețete",
        "description": "Căutare rețete după nume sau tip",
        "parameters": [
          {
            "in": "path",
            "name": "query",
            "required": true,
            "schema": { "type": "string" },
            "description": "Termenul de căutare"
          }
        ],
        "responses": {
          "200": {
            "description": "Lista rețetelor găsite",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/Recipe" }
                }
              }
            }
          },
          "500": { "description": "Eroare server" }
        }
      }
    },
    "/api/recipes/cuisines": {
      "get": {
        "summary": "Preia toate bucătăriile",
        "description": "Preia lista tuturor bucătăriilor din baza de date",
        "responses": {
          "200": {
            "description": "Lista tuturor bucătăriilor",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/Cuisine" }
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Adaugă o bucătărie nouă",
        "description": "Creează o bucătărie nouă în baza de date",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": { "type": "string" },
                  "description": { "type": "string" },
                  "image": { "type": "string" }
                },
                "required": ["name"]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Bucătăria a fost creată cu succes",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Cuisine" }
              }
            }
          },
          "500": { "description": "Eroare server" }
        }
      }
    },
    "/api/recipes/cuisines/{id}": {
      "delete": {
        "summary": "Șterge o bucătărie",
        "description": "Șterge o bucătărie din baza de date",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": { "type": "integer" },
            "description": "ID-ul bucătăriei"
          }
        ],
        "responses": {
          "200": { "description": "Bucătăria a fost ștearsă cu succes" },
          "500": { "description": "Eroare server" }
        }
      }
    },
    "/api/recipes/cuisines/{id}/image": {
      "patch": {
        "summary": "Actualizează imaginea bucătăriei",
        "description": "Actualizează imaginea unei bucătării existente",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": { "type": "integer" },
            "description": "ID-ul bucătăriei"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "image": { "type": "string" }
                },
                "required": ["image"]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Imaginea bucătăriei a fost actualizată cu succes",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Cuisine" }
              }
            }
          },
          "500": { "description": "Eroare server" }
        }
      }
    },
    "/api/recipes/cuisine/{cuisine}": {
      "get": {
        "summary": "Preia rețetele după bucătărie",
        "description": "Filtrarea rețetelor în funcție de tipul de bucătărie",
        "parameters": [
          {
            "in": "path",
            "name": "cuisine",
            "required": true,
            "schema": { "type": "string" },
            "description": "Tipul bucătăriei (ex: romanian, italian, american)"
          }
        ],
        "responses": {
          "200": {
            "description": "Lista rețetelor după tipul de bucătărie",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/Recipe" }
                }
              }
            }
          }
        }
      }
    },
...
  }
}